pyproject.toml
[project]
name = "sql-dialect-mapper"
version = "0.2.0"
description = "SQL dialect function mapping with mathematical basis synthesis"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "deepagents>=0.1.0",
    "pyspark>=3.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dialect_mapper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
